refactor(LadderRowDto): 레이어 간 데이터 교환을 목적으로 하는 클래스명을 `Dto`로 끝나도록 하고, 패키지 분리
